Added a title screen
I added to the Goal Manager script, a clause which allows player 1 to
press a on the title screen, and it will load the first screen. Once we
add more levels to the build in order, we will be able to load the more
levels as we go through the game. Testing that functionality is next.